+++ src/app/cart/checkout/order-info/order-info.module.scss
@@ -248,129 +248,3 @@
     text-decoration: underline;
   }
 }
-
-.asidePanel {
-  margin-top: 48px;
-  background-color: $gray-80;
-  border: 1px solid $gray-10;
-  color: $gray-0;
-  border-radius: 24px;
-  padding: 24px;
-
-  @include laptop {
-    @include set-flex(column);
-
-    position: sticky;
-    top: 20px;
-    margin-top: 24px;
-    width: 35%;
-    height: min-content;
-  }
-
-  &__title {
-    color: $gray-0;
-    font-size: $font-size-lg;
-  }
-
-  &__shippingCost {
-    margin-top: 24px;
-    margin-bottom: -24px;
-    color: $gray-10;
-    font-size: $font-size-xl;
-    font-family: $font-title;
-  }
-
-  &__totalPrice {
-    margin-top: 24px;
-    font-size: $font-size-xxl;
-    font-family: $font-title;
-  }
-}
-
-.orderItem {
-  &__paintingWrapper {
-    @include set-flex(row);
-
-    margin-top: 24px;
-    gap: 16px;
-  }
-
-  &__imageWrapper {
-    @include set-square(80px);
-
-    position: relative;
-
-    @include desktop {
-      @include set-square(100px);
-    }
-  }
-
-  &__paintingInfo {
-    @include set-flex(column, space-between);
-
-    flex: 1;
-    width: 100%;
-  }
-
-  &__title {
-    font-size: $font-size-base;
-  }
-
-  &__author {
-    font-size: $font-size-md;
-    color: $gray-10;
-  }
-
-  &__price {
-    @include set-flex(row, space-between);
-  }
-}
-
-.saveOrder {
-  &__secure {
-    @include set-flex(row);
-
-    margin-top: 24px;
-    gap: 24px;
-  }
-
-  &__info {
-    flex: 1;
-  }
-
-  &__title {
-    font-size: $font-size-lg;
-    color: $gray-5;
-  }
-
-  &__text {
-    margin-top: 16px;
-    color: $gray-10;
-  }
-}
-
-.button {
-  @include set-flex(row, center, center);
-
-  margin-top: 48px;
-  width: 100%;
-  height: 48px;
-  border-radius: 24px;
-  background-color: $gray-5;
-  color: $gray-100;
-  cursor: pointer;
-
-  &:hover {
-    background-color: $gray-0;
-  }
-
-  &:active {
-    transform: scale(0.9);
-    transition-duration: 0.3s;
-  }
-}
-
-.errorInfo {
-  margin-top: 16px;
-  color: $system-error;
-}
