+++ src/app/components/button-loader/button-loader.module.scss
@@ -1,0 +1,45 @@
+@import "@styles/mixins.scss";
+@import "@styles/vars.scss";
+
+.wrapper {
+  @include set-flex(column, center, center);
+
+  z-index: 99999;
+}
+
+.loader {
+  @include set-square(30px);
+
+  border-radius: 50%;
+  display: inline-block;
+  border-top: 4px solid #fff;
+  border-right: 4px solid transparent;
+  box-sizing: border-box;
+  animation: rotation 1s linear infinite;
+}
+
+.darkLoader {
+  border-top: 4px solid $gray-80;
+}
+
+.loader::after {
+  @include set-square(30px);
+
+  content: "";
+  box-sizing: border-box;
+  position: absolute;
+  left: 0;
+  top: 0;
+  border-radius: 50%;
+  border-left: 4px solid #ff3d00;
+  border-bottom: 4px solid transparent;
+  animation: rotation 0.5s linear infinite reverse;
+}
+@keyframes rotation {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
