+++ src/app/components/button-loader/button-loader.tsx
@@ -1,0 +1,13 @@
+import style from "./button-loader.module.scss";
+
+const ButtonLoader = ({ darkLoader = false }: { darkLoader?: boolean }) => {
+  return (
+    <div className={style.wrapper}>
+      <span
+        className={`${style.loader} ${darkLoader && style.darkLoader}`}
+      ></span>
+    </div>
+  );
+};
+
+export default ButtonLoader;
