+++ src/utils/api.ts
@@ -227,7 +227,7 @@ export async function updateProfile(userData: UserDataToSave, headers: object) {
   return data;
 }
 
-export async function saveOrderPaintingToServer(id: any, headers: object) {
+export async function saveOrderPaintingToServer(id: number, headers: object) {
   const { data } = await axios.get(`${BASE_URL}cart/add/${id}`, {
     headers,
   });
@@ -243,7 +243,7 @@ export async function saveOrderPaintingsToServer(ids: string, headers: object) {
   return data;
 }
 
-export async function removeOrderPaintingFromServer(id: any, headers: object) {
+export async function removeOrderPaintingFromServer(id: number, headers: object) {
   const { data } = await axios.get(`${BASE_URL}cart/remove/${id}`, {
     headers,
   });
